+++ addOns/automation/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Support for dynamically added auto-detect authentication.
+- Support for dynamically added auto-detect session management.
+
 ### Fixed
 - Correct file path resolution with environment variables.